app/requirements.txt
django==2.2.10
django-filter==21.1
djangorestframework==3.11.2
djangorestframework-simplejwt==5.0.0
drf-yasg
psycopg2-binary==2.8.6
requests==2.26.0

